moved explainability from core to API;
