+++ examples/fedot_api_example.py
@@ -101,6 +101,11 @@ def run_classification_multiobj_example(with_plot=True, timeout=None):
     return prediction
 
 
+def run_explain_example():
+    # TODO explainability: API explain example
+    pass
+
+
 if __name__ == '__main__':
     run_classification_example()
     run_regression_example()
